separate simulator and cli
