+++ cli/go.mod
@@ -1,0 +1,11 @@
+module ubiwhere
+
+go 1.14
+
+require (
+	github.com/gin-gonic/gin v1.6.3
+	github.com/jinzhu/gorm v1.9.12
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/spf13/cobra v1.0.0
+	github.com/spf13/viper v1.7.0
+)
