+++ simulator/go.mod
@@ -1,0 +1,22 @@
+module ubiwhere
+
+go 1.14
+
+require (
+	github.com/fsnotify/fsnotify v1.4.9 // indirect
+	github.com/gin-gonic/gin v1.6.3
+	github.com/jinzhu/gorm v1.9.12
+	github.com/mackerelio/go-osstat v0.1.0
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/mitchellh/mapstructure v1.3.0 // indirect
+	github.com/pelletier/go-toml v1.8.0 // indirect
+	github.com/spf13/afero v1.2.2 // indirect
+	github.com/spf13/cast v1.3.1 // indirect
+	github.com/spf13/cobra v1.0.0
+	github.com/spf13/jwalterweatherman v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/spf13/viper v1.7.0
+	github.com/stretchr/testify v1.5.1
+	golang.org/x/sys v0.0.0-20200515095857-1151b9dac4a9 // indirect
+	gopkg.in/ini.v1 v1.56.0 // indirect
+)
